):feat(group 优化群组加入逻辑并添加退出群组功能Group 方法，实现了用户退出群组- 修改了 GroupController 中的 out的功能 - 优化了 GroupJoinController 中的 joinGroup 方法，增加了对重复申请的处理 - 更新了 GroupJoinServerImpl 中的 agreeJoin 方法，确保只有学生用户才会增加群组人数 -调整了 User 类和 UserDTO 类中的 id 类型，从 int 改为 Long - 更新了 application.yml 中的文件上传 URL前缀

+++ studyParty_entityApi/src/main/java/com/studyParty/entity/user/User.java
@@ -21,7 +21,7 @@ public class User implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
     @TableId(type = IdType.AUTO)
-    private int id;
+    private Long id;
     private String name;
     private String password;
     private String head;
@@ -63,7 +63,7 @@ public class User implements Serializable {
         this.finishTask = 0;
     }
     /// 修改个人信息专用
-    public User(int id, String name, String sex, String major, String grade, String phone, String school, String email) {
+    public User(Long id, String name, String sex, String major, String grade, String phone, String school, String email) {
         this.id = id;
         this.name = name;
         this.sex = sex;
